requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.1
python-dotenv==1.0.1
groq==0.9.0
sqlalchemy==2.0.31
